[cheri][shared][drivers]: Add init_console
UART drivers directly cast integer addresses to perform MMIO operations.
This does not work with CHERI C as all memory accesses need to be
performed via valid capabilities.

This commit introduces init_console() to be called at boot so that CHERI
boostraping code can build valid capabilities for it. Currently, it's
only Morello's PL011 driver that is ported to CHERI.

Apart from CHERI, seL4 could also use these functions in the future, if
needed, to actually initialise the UART devices without relying on
bootloaders doing that for it.

Signed-off-by: Hesham Almatary <heshamalmatary@capabilitieslimited.co.uk>