[cheri][riscv] Add a new Toooba-BESSPIN platform
This commit adds support for Toooba: a CHERI-Enabled Out-of-Order RISC-V Core.
It can run both unmodified RISC-V code as well as CHERI-RISC-V Cambridge
ISAv9 [2].

BESSPIN is the SoC name and it incorporates Toooba as a CHERI-RISC-V core and
runs on FPGAs. The port has been tested and validated on VCU-118 board passing
sel4tests, with the following variants:
1- Vanilla RISC-V (non-CHERI).
2- Hybrid kernel, purecap userspace (sel4test)
3- Purecap kernel and userspace (sel4test)

[1] https://github.com/CTSRD-CHERI/Toooba
[2] https://www.cl.cam.ac.uk/techreports/UCAM-CL-TR-987.pdf
[3] https://github.com/CTSRD-CHERI/BESSPIN-GFE/commit/d3dcba9d1bef3fd373e0e973a00a92e201f14cde

Signed-off-by: Hesham Almatary <heshamalmatary@capabilitieslimited.co.uk>